Fix rigctld dump_state: remove v1 extended fields, fix DAX TX routing
dump_state:
- Removed 14 extra v1 protocol fields (vfo_ops, ptt_type, targetable_vfo,
  has_set/get capability flags, timeout, done) that real rigctld model 2
  does not send. WSJT-X was reading these as corrupted command responses,
  causing rapid connect/disconnect cycles.
- Now matches exact output of Hamlib 4.6.5 rigctld --model=2

DAX TX:
- Mute mic capture (QAudioSource) when DAX enabled to prevent mic audio
  from competing with DAX TX virtual sink audio on the same VITA-49 stream
- Wire daxTxDataReady signal to feedDaxTxData for VITA-49 packet encoding
- PipeWire handles 24k↔48k resampling (no manual up/downsample needed)
- Manual pw-link for TX capture stream (AUTOCONNECT targets wrong node)
- Debug logging every 100th DAX TX packet

Rigctld logging:
- Suppress single-char polling commands (v/f/m/t) to reduce log spam

+++ src/core/PanadapterStream.h
@@ -149,7 +149,7 @@ private:
     float           m_maxDbm{-20.0f};
     RadioConnection* m_conn{nullptr};
     FrameAssembler  m_frame;
-    QMap<quint16, StreamStats> m_streamStats;  // keyed by PCC
+    QMap<quint32, StreamStats> m_streamStats;  // keyed by stream ID
 
 public:
     // Packet error/total counts across all streams (for network quality monitor).
